refactor: restructure compare as feature slice
