+++ src/components/organisms/CompareModal/ResourceSetSelector/ClusterContextSelect.tsx
@@ -4,8 +4,8 @@ import {Select} from 'antd';
 
 import invariant from 'tiny-invariant';
 
+import {PartialResourceSet, resourceSetSelected, selectClusterResourceSet} from '@redux/compare';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {PartialResourceSet, resourceSetSelected, selectClusterResourceSet} from '@redux/reducers/compare';
 
 type Props = {
   side: 'left' | 'right';
