+++ src/components/organisms/CompareModal/TransferButton.tsx
@@ -6,8 +6,8 @@ import {ArrowLeftOutlined, ArrowRightOutlined} from '@ant-design/icons';
 
 import invariant from 'tiny-invariant';
 
+import {CompareSide, selectCanTransfer, selectCompareStatus, transferResource} from '@redux/compare';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
-import {CompareSide, selectCanTransfer, selectCompareStatus, transferResource} from '@redux/reducers/compare';
 
 type Props = {
   side: CompareSide;
