+++ src/redux/compare/slice.ts
@@ -1,0 +1,272 @@
+import {PayloadAction, createSlice} from '@reduxjs/toolkit';
+
+import {WritableDraft} from 'immer/dist/internal';
+import log from 'loglevel';
+
+import {K8sResource} from '@models/k8sresource';
+
+import {trackEvent} from '@utils/telemetry';
+
+import {selectIsAllComparisonSelected} from './selectors';
+import {
+  CompareFilter,
+  CompareOperation,
+  CompareSide,
+  CompareState,
+  ComparisonInspection,
+  ComparisonView,
+  PartialResourceSet,
+  ResourceComparison,
+  initialState,
+} from './state';
+import {transferResource} from './thunks';
+
+export const compareSlice = createSlice({
+  name: 'compare',
+  initialState,
+  reducers: {
+    compareToggled: (
+      state,
+      action: PayloadAction<{
+        value: boolean | undefined;
+        initialView?: ComparisonView;
+        from?: 'compare-button' | 'quick-helm-compare' | 'quick-kustomize-compare';
+      }>
+    ) => {
+      const {value, initialView} = action.payload;
+      const close = value === undefined ? state.isOpen : value === false;
+
+      if (close) {
+        return initialState;
+      }
+
+      state.isOpen = true;
+      if (initialView) {
+        state.current.view = initialView;
+      }
+
+      trackEvent('COMPARE_OPENED', {from: action.payload.from});
+    },
+    operationUpdated: (state, action: PayloadAction<{operation: CompareOperation}>) => {
+      state.current.view.operation = action.payload.operation;
+      resetComparison(state);
+    },
+    searchUpdated: (state, action: PayloadAction<{search: string | undefined}>) => {
+      state.current.search = action.payload.search;
+    },
+    namespaceUpdated: (state, action: PayloadAction<{namespace: string | undefined}>) => {
+      state.current.view.namespace = action.payload.namespace;
+    },
+    filterUpdated: (state, action: PayloadAction<{filter: CompareFilter | undefined}>) => {
+      const newFilter = action.payload.filter;
+      const isEmpty = !newFilter || Object.keys(newFilter).length === 0;
+
+      if (isEmpty) {
+        state.current.view.filter = undefined;
+      } else {
+        state.current.view.filter = action.payload.filter;
+      }
+    },
+    resourceSetSelected: (state, action: PayloadAction<{side: CompareSide; value: PartialResourceSet}>) => {
+      const {side, value} = action.payload;
+      resetComparison(state);
+      if (side === 'left') {
+        state.current.view.leftSet = value;
+        state.current.left = undefined;
+      } else {
+        state.current.view.rightSet = value;
+        state.current.right = undefined;
+      }
+    },
+    resourceSetCleared: (state, action: PayloadAction<{side: CompareSide | 'both'}>) => {
+      const side = action.payload.side;
+      resetComparison(state);
+
+      if (['left', 'both'].includes(side)) {
+        state.current.view.leftSet = undefined;
+        state.current.left = undefined;
+      }
+      if (['right', 'both'].includes(side)) {
+        state.current.view.rightSet = undefined;
+        state.current.right = undefined;
+      }
+    },
+    resourceSetRefreshed: (state, action: PayloadAction<{side: CompareSide}>) => {
+      const {side} = action.payload;
+      resetComparison(state);
+      state.current[side] = undefined;
+    },
+    comparisonInspecting: (state, action: PayloadAction<ComparisonInspection>) => {
+      trackEvent('COMPARE_INSPECTED', {type: state.current.inspect?.type});
+      state.current.inspect = action.payload;
+    },
+    comparisonInspected: state => {
+      state.current.inspect = undefined;
+    },
+    comparisonToggled: (state, action: PayloadAction<{id: string}>) => {
+      const {id} = action.payload;
+
+      const index = state.current.selection.findIndex(comparison => comparison === id);
+      if (index === -1) {
+        state.current.selection.push(id);
+      } else {
+        state.current.selection.splice(index, 1);
+      }
+    },
+    comparisonAllToggled: state => {
+      const isAllSelected = selectIsAllComparisonSelected(state);
+      if (isAllSelected) {
+        state.current.selection = [];
+      } else {
+        state.current.selection = state.current.comparison?.comparisons.map(c => c.id) ?? [];
+      }
+    },
+    resourceSetFetchPending: (state, action: PayloadAction<{side: CompareSide}>) => {
+      const {side} = action.payload;
+
+      state.current[side] = {
+        loading: true,
+        error: false,
+        resources: [],
+      };
+    },
+    resourceSetFetched: (state, action: PayloadAction<{side: CompareSide; resources: K8sResource[]}>) => {
+      const {side, resources} = action.payload;
+      state.current[side] = {
+        error: false,
+        loading: false,
+        resources,
+      };
+
+      const leftReady = state.current.left && !state.current.left.loading && !state.current.left.error;
+      const rightReady = state.current.right && !state.current.right.loading && !state.current.right.error;
+      if (leftReady && rightReady) {
+        state.current.comparison = {
+          loading: true,
+          comparisons: [],
+        };
+      }
+    },
+    resourceSetFetchFailed: (state, action: PayloadAction<{side: CompareSide; reason: string}>) => {
+      const {side, reason} = action.payload;
+      log.debug('[compare] resource set fetch failed', reason);
+
+      state.current[side] = {
+        error: true,
+        loading: false,
+        resources: [],
+      };
+    },
+    resourceSetCompared: (state, action: PayloadAction<{comparisons: ResourceComparison[]}>) => {
+      trackEvent('COMPARE_COMPARED', {
+        left: state.current.view.leftSet?.type,
+        right: state.current.view.rightSet?.type,
+        operation: state.current.view.operation ?? 'default',
+      });
+
+      state.current.comparison = {
+        loading: false,
+        comparisons: action.payload.comparisons,
+      };
+    },
+    resourceSetFiltered: (state, action: PayloadAction<{comparisons: ResourceComparison[]}>) => {
+      state.current.filtering = {
+        comparisons: action.payload.comparisons,
+      };
+    },
+  },
+  extraReducers: builder => {
+    builder.addCase(transferResource.pending, state => {
+      state.current.transfering.pending = true;
+    });
+    builder.addCase(transferResource.rejected, state => {
+      state.current.transfering.pending = false;
+    });
+    builder.addCase(transferResource.fulfilled, (state, action) => {
+      const {side, delta} = action.payload;
+
+      const direction = action.meta.arg.direction;
+      const leftSet = state.current.view.leftSet;
+      const rightSet = state.current.view.rightSet;
+      const from = direction === 'left-to-right' ? leftSet?.type : rightSet?.type;
+      const to = direction === 'left-to-right' ? rightSet?.type : leftSet?.type;
+      trackEvent('COMPARE_TRANSFERED', {from, to, count: delta.length});
+
+      state.current.transfering.pending = false;
+
+      // Splice each update into previous results to give smooth UX.
+      delta.forEach(comparison => {
+        // Remove from selection
+        state.current.selection = state.current.selection.filter(s => s !== comparison.id);
+
+        // Update comparison without recomputing
+        if (state.current.comparison) {
+          const index = state.current.comparison.comparisons.findIndex(c => c.id === comparison.id) ?? -1;
+          if (index !== -1) {
+            state.current.comparison.comparisons.splice(index, 1, comparison);
+          } else {
+            state.current.comparison.comparisons.push(comparison);
+          }
+        }
+
+        // Update filtering without recomparing
+        if (state.current.filtering) {
+          const index = state.current.filtering.comparisons.findIndex(c => c.id === comparison.id) ?? -1;
+          if (index !== -1) {
+            state.current.filtering.comparisons.splice(index, 1, comparison);
+          } else {
+            state.current.filtering.comparisons.push(comparison);
+          }
+        }
+
+        // Update resource sets without refetching
+        if (side === 'left' && state.current.left) {
+          const index = state.current.left.resources.findIndex(r => r.id === comparison.left.id) ?? -1;
+          if (index === -1) {
+            state.current.left.resources.push(comparison.left);
+          } else {
+            state.current.left.resources[index] = comparison.left;
+          }
+        }
+        if (side === 'right' && state.current.right) {
+          const index = state.current.right.resources.findIndex(r => r.id === comparison.right.id) ?? -1;
+          if (index === -1) {
+            state.current.right.resources.push(comparison.right);
+          } else {
+            state.current.right.resources[index] = comparison.right;
+          }
+        }
+      });
+    });
+  },
+});
+
+function resetComparison(state: WritableDraft<CompareState>) {
+  state.current.comparison = undefined;
+  state.current.inspect = undefined;
+  state.current.selection = [];
+
+  state.current.filtering = undefined;
+}
+
+export const {
+  compareToggled,
+  searchUpdated,
+  namespaceUpdated,
+  filterUpdated,
+  operationUpdated,
+  resourceSetCleared,
+  resourceSetRefreshed,
+  resourceSetSelected,
+  resourceSetFetchPending,
+  resourceSetFetchFailed,
+  resourceSetFetched,
+  resourceSetCompared,
+  resourceSetFiltered,
+  comparisonInspecting,
+  comparisonInspected,
+  comparisonAllToggled,
+  comparisonToggled,
+} = compareSlice.actions;
+
+export default compareSlice.reducer;
