+++ src/redux/compare/utils.ts
@@ -1,0 +1,20 @@
+import {isDefined} from '@utils/filter';
+
+import {PartialResourceSet, ResourceSet} from './state';
+
+export function isCompleteResourceSet(options: PartialResourceSet | undefined): options is ResourceSet {
+  switch (options?.type) {
+    case 'local':
+      return true;
+    case 'cluster':
+      return isDefined(options.context);
+    case 'kustomize':
+      return isDefined(options.kustomizationId);
+    case 'helm':
+      return isDefined(options.chartId) && isDefined(options.valuesId);
+    case 'helm-custom':
+      return isDefined(options.chartId) && isDefined(options.configId);
+    default:
+      return false;
+  }
+}
