+++ src/redux/services/compare/transferResource.ts
@@ -7,7 +7,7 @@ import {AppDispatch} from '@models/appdispatch';
 import {K8sResource} from '@models/k8sresource';
 import {RootState} from '@models/rootstate';
 
-import {ResourceSet} from '@redux/reducers/compare';
+import {ResourceSet} from '@redux/compare';
 import {currentConfigSelector, kubeConfigContextSelector, kubeConfigPathSelector} from '@redux/selectors';
 import {applyResource} from '@redux/thunks/applyResource';
 import {updateResource} from '@redux/thunks/updateResource';
